+++ app-common-io/src/main/aidl/eu/darken/sdmse/common/shell/ipc/ShellOpsCmd.aidl
@@ -1,0 +1,5 @@
+package eu.darken.sdmse.common.shell.ipc;
+
+import eu.darken.sdmse.common.shell.ipc.ShellOpsCmd;
+
+parcelable ShellOpsCmd;
