+++ app-common-io/src/main/aidl/eu/darken/sdmse/common/shell/ipc/ShellOpsConnection.aidl
@@ -1,0 +1,10 @@
+package eu.darken.sdmse.common.shell.ipc;
+
+import eu.darken.sdmse.common.shell.ipc.ShellOpsCmd;
+import eu.darken.sdmse.common.shell.ipc.ShellOpsResult;
+
+interface ShellOpsConnection {
+
+   ShellOpsResult execute(in ShellOpsCmd cmd);
+
+}
