+++ app-common-io/src/main/aidl/eu/darken/sdmse/common/shell/ipc/ShellOpsResult.aidl
@@ -1,0 +1,5 @@
+package eu.darken.sdmse.common.shell.ipc;
+
+import eu.darken.sdmse.common.shell.ipc.ShellOpsResult;
+
+parcelable ShellOpsResult;
