--- app-common-io/src/main/aidl/eu/darken/sdmse/common/shell/root/ShellOpsConnection.aidl
@@ -1,10 +1,0 @@
-package eu.darken.sdmse.common.shell.root;
-
-import eu.darken.sdmse.common.shell.root.ShellOpsCmd;
-import eu.darken.sdmse.common.shell.root.ShellOpsResult;
-
-interface ShellOpsConnection {
-
-   ShellOpsResult execute(in ShellOpsCmd cmd);
-
-}
