--- app-common-io/src/main/aidl/eu/darken/sdmse/common/shell/root/ShellOpsResult.aidl
@@ -1,5 +1,0 @@
-package eu.darken.sdmse.common.shell.root;
-
-import eu.darken.sdmse.common.shell.root.ShellOpsResult;
-
-parcelable ShellOpsResult;
