+++ app-common-io/src/main/aidl/eu/darken/sdmse/common/shizuku/ShizukuServiceConnection.aidl
@@ -2,7 +2,7 @@ package eu.darken.sdmse.common.shizuku;
 
 import eu.darken.sdmse.common.files.local.ipc.FileOpsConnection;
 import eu.darken.sdmse.common.pkgs.pkgops.ipc.PkgOpsConnection;
-import eu.darken.sdmse.common.shell.root.ShellOpsConnection;
+import eu.darken.sdmse.common.shell.ipc.ShellOpsConnection;
 
 interface ShizukuServiceConnection {
     String checkBase();
