+++ app-common-io/src/main/java/eu/darken/sdmse/common/shell/ipc/ShellOpsResult.kt
@@ -1,4 +1,4 @@
-package eu.darken.sdmse.common.shell.root
+package eu.darken.sdmse.common.shell.ipc
 
 import android.os.Parcelable
 import eu.darken.rxshell.cmd.Cmd
