+++ app-common-pkgs/src/main/java/eu/darken/sdmse/common/pkgs/sources/ShellPkgSource.kt
@@ -17,7 +17,7 @@ import eu.darken.sdmse.common.pkgs.pkgops.PkgOps
 import eu.darken.sdmse.common.root.RootManager
 import eu.darken.sdmse.common.root.useRootNow
 import eu.darken.sdmse.common.shell.ShellOps
-import eu.darken.sdmse.common.shell.root.ShellOpsCmd
+import eu.darken.sdmse.common.shell.ipc.ShellOpsCmd
 import eu.darken.sdmse.common.user.UserManager2
 import java.util.*
 import javax.inject.Inject
